refactor(auth): extract JWT and refresh token services to shared module
- Move JwtTokenService to shared/application/services for reusability across modules
- Extract RefreshTokenEntity and repository to shared/domain layer
- Move RefreshTokenPrismaAdapter to shared/infrastructure/adapters
- Add EventBusService for domain event handling
- Add PrismaErrorHandlerService for centralized error handling
- Create GraphQLExceptionFilter for global exception handling
- Move JWT types to shared/domain/types for better organization
- Simplify AuthModule by removing complex provider configurations
- Update app.module.ts with jwtConfig and correct filter import

Follows hexagonal architecture principles with clear separation between
domain, application, and infrastructure layers. Improves code reusability
and maintainability by moving common JWT functionality to shared module.
