+++ src/modules/auth/application/services/index.ts
@@ -1,4 +1,2 @@
 // Domain layer services barrel exports
 // Import and export all domain services here as they are implemented
-
-export * from './refresh-token-domain.service';
