+++ src/modules/auth/domain/repositories/index.ts
@@ -2,5 +2,3 @@
 // Import and export all repository interfaces here as they are implemented
 
 export * from './auth.repository';
-export * from './refresh-token.repository';
-export * from './token.repository';
