+++ src/modules/auth/domain/types/auth.types.ts
@@ -1,14 +1,3 @@
-import { UserEntity } from '@/modules/user/domain/entities';
-
-/**
- * Enum defining different types of temporary JWT tokens used for various authentication flows
- */
-export enum JwtTempTokenType {
-  FORGOT_PASSWORD = 'forgot_password',
-  RESET_PASSWORD = 'reset_password',
-  REFRESH_TOKEN = 'refresh_token',
-}
-
 /**
  * Authentication credentials for user sign-in
  * Supports both email and username for flexible authentication
@@ -30,31 +19,6 @@ export interface UserRegistrationData {
   readonly userName: string;
   /** Plain text password that will be hashed */
   readonly password: string;
-}
-
-/**
- * JWT token payload structure containing user information
- */
-export interface JwtPayload {
-  /** User's unique identifier (user ID) */
-  sub: string;
-  /** Complete user entity with all user information */
-  user: UserEntity;
-  /** Token issued at timestamp */
-  iat?: number;
-  /** Token expiration timestamp */
-  exp?: number;
-}
-
-/**
- * Temporary token payload structure for specific actions
- * Uses a unique identifier (UUID) instead of user ID for database validation
- */
-export interface TempTokenPayload extends Omit<JwtPayload, 'sub'> {
-  /** Unique token identifier (UUID) for database validation */
-  sub: string;
-  /** Type of temporary token */
-  type: JwtTempTokenType;
 }
 
 /**
