--- src/modules/auth/infrastructure/adapters/index.ts
@@ -1,5 +1,0 @@
-// Infrastructure layer adapters barrel exports
-// Import and export all adapters (repository implementations) here as they are implemented
-
-export * from './jwt-token.adapter';
-export * from './refresh-token-prisma.adapter';
