+++ src/modules/auth/infrastructure/decorators/index.ts
@@ -1,4 +1,2 @@
 // Infrastructure layer decorators barrel exports
 // Import and export all decorators here as they are implemented
-
-export * from './public.decorator';
