--- src/modules/auth/infrastructure/guards/index.ts
@@ -1,4 +1,0 @@
-// Infrastructure layer guards barrel exports
-// Import and export all guards here as they are implemented
-
-export * from './auth.guard';
