+++ src/shared/application/filters/graphql-exception.filter.ts
