+++ src/shared/application/services/event-bus.service.ts
