--- src/shared/application/services/events/index.ts
@@ -1,1 +1,0 @@
-export * from './event-bus.service';
