+++ src/shared/application/services/index.ts
@@ -1,2 +1,4 @@
+export * from './event-bus.service';
+export * from './jwt-token.service';
+export * from './prisma-error-handler.service';
 export * from './prisma.service';
-export * from './events';
