+++ src/shared/domain/entities/index.ts
