+++ src/shared/domain/repositories/index.ts
@@ -1,0 +1,1 @@
+export * from './refresh-token.repository';
