+++ src/shared/domain/types/index.ts
@@ -1,1 +1,2 @@
 export * from './pagination.types';
+export * from './jwt.types';
