+++ src/shared/infrastructure/adapters/index.ts
@@ -1,0 +1,1 @@
+export * from './refresh-token-prisma.adapter';
