--- src/shared/infrastructure/services/index.ts
@@ -1,1 +1,0 @@
-export * from './prisma-error-handler.service';
